Use make_config to configure interfaces and bridges in Nm
Nm used to make call to the low-level networkd functions to create
and configure bridges and interfaces when "bringing up" a PIF.

Now, Nm essentially transforms the relevant bit of the xapi database
into a networkd configuration structure, which is passed into
Bridge.make_config and Interface.make_config. These two networkd
functions take care of setting things up in the right order, using the
same code that is used when the hosts starts up.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>