Allow Nm.bring_pif_up (down) even when currently_attached is not false (true)
This is possible, because all functions of networkd are idempotent.

PIF.plug can now be used to sync networkd's internal state, even when
a PIF is already up, to ensure that it come back correctly after a reboot.

Furthermore, it is now no longer necessary to unplug a PIF when changing
a parameter such as the MTU. To change the MTU, set the new value on the
network object, and then do a PIF.plug right away to make it happen.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>