CA-73892: networkd: after upgrade, use the old script to bring up NICs
This needs to happen before xapi starts (especially on a pool slave).
Networkd does not have any state yet, because it was not present before
the upgrade. Therefore, it uses the old init script (management-interface)
to initially bring up and configure all interfaces.

When xapi starts up after that, it will sync up with networkd through
the "Best-effort bring up of physical NICs" startup item, in order to
fill networkd's cache.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>