CP-3477: Make xapi listen on all dom0 IP addresses
Xapi now starts listening on all IP addresses when it starts up, and
management is not disabled.

Even though xapi listens on all IP addresses, there is still an interface
appointed as _the_ management interface. Slaves in a pool use the IP address
of this interface to connect the pool master.

It is now no longer needed to rebind the management socket after IP changes.

Changing the management interface with Host.management_reconfigure now starts
the API server in case it was not running, but leaves it alone otherwise.
Besides that, it is only an update of the inventory file, plus an update of the
Host.address field.

Also remove the rewrite-management-interface script, and instead update the
MANAGEMENT_INTERFACE key in inventory file in xapi itself.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>