Use HTTP "host" header to decided whether to forward and API call to the master
Since xapi is now listening on all dom0 IPs, API calls coming in on
the HIMN now go through the API server as all other incoming calls.

The different between calls on the HIMN and those coming from outside
is that HIMN calls on a pool slave must be forwarded to the master, while
others do not. We use the HTTP "host" header to distinguish between these
cases.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>