CA-77823: Add argument for network choice when moving a VM.
The network can be specified via the network parameter of
Host.migrate_receive, or the remote-network parameter on the CLI. If the
network is not specified on the CLI, xapi will attempt to find the host
internal management network on the remote host.

n.b. The network argument does not yet have any effect on the actual
migration.

Signed-off-by: John Else <john.else@citrix.com>