IPv6 management: XAPI management interface on IPv6
XAPI can now listen on either IPv4 or IPv6 addresses; the "primary
address type" of the management interface defines which one of the
two it is.

You can only change management to a PIF that has:
1. the same primary address type as the current management PIF, and
2. IP configuration of the required type (i.e. not "none").

You therefore need to make the choice for IPv4 or IPv6 management
from the start, as the primary address type can be changed only by
first disabling management.

Signed-off-by: Bob Ball <bob.ball@citrix.com>
Signed-off-by: Rob Hoes <rob.hoes@citrix.com>