CA-76600: preserve /vm and /vss over localhost migrate
Rather than having the higher-level software special case localhost
migrate, we use reference counting to decide when to delete /vm
and /vss. When we create a new domain for a specific UUID, we
record the domain in
  /vm/<uuid>/domains/<domid> = <local domain path>

In domain destroy, we delete /vm and /vss if the domains directory
is empty.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>