CA-77555: add exception handlers around each xenopsd VM event stanza, so one failure does not block the rest
It's important that the event resync logic doesn't fail, otherwise the VM state in xapi will be out-of-sync.

In the event that something does fail, log the error and resynchronise as much as possible.

In future we may wish to consider artificially injecting events periodically in the event of failure, in the hope that the state will change enough in future to avoid the problem.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>