IPv6 management: monitor IPv6 address changes in networkd and signal xapi if needed
Uses "ip monitor address" from iproute2 to track updates on IPv6 addresses.

This serves the same purpose as the calls to 'xe host-signal-networking-change' that
are done by dhclient-script in the IPv4 case. IPv6 is slightly more complex, because besides
by DHCP, addresses may be added by prefix delegation ("autoconf"), for which there is no
callback.

If this works well, we may consider doing it for IPv4 as well. We can then leave dhclient-script
unmodified (good for Kronos etc).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>