CP-3482: serialise the device create xenstore transaction, to work around an inefficiency in oxenstored
The oxenstored transaction merge/conflict algorithm will merge two transactions if the lowest nodes which are common to all the paths modified in a transaction are in separate subtrees. Unfortunately a device create adds both a backend and a frontend in different domains, so the lowest common node is always /local, causing an abort and leading to livelock.

We work around this for now by serialising the device create operations in xenopsd.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>