Move 'special' xenstore setup into appropriate init.d scripts
Anomalously, domain 0 is the only domain which doesn't run the guest
agent -- we work around this by writing the feature-balloon flag
in the squeezed init script, since squeezed is the only one who cares.

We still have a special VBD shutdown protocol for domain 0, so
write this "feature" flag in the xenopsd init script, since squeezed
is the only one who cares.

Xapi's init.d script should not touch xenstore because xapi should
not touch xenstore.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>