Move the setup of domain 0's minimal xenstore keys to xenopsd
Domain 0 is the only current example of a pre-built domain which
is to be managed by xenopsd. Tools such as xentop and even xenopsd
itself require certain xenstore keys to be present ("name" for
xentop and "vm" for xenopsd) for proper functioning. Therefore
if a VM is registered with xenopsd via VM.add we check if any
keys are missing and add them if so.

This should fix starting VMs on Debian.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>