Change the way the default gateway and DNS interfaces are set
Xapi now explicitly tells networkd which interfaces are the gateway and DNS
interfaces. Bringing up or reconfiguring an interface will only cause
the gateway or DNS servers to be set if it is the gateway or DNS interface
respectively. This applies to interfaces configured by DHCP as well as those
with static addresses.

The patch also removes the "options" from the DHCP type in the
Network_interface, since it has no meaning anymore.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>