CA-73099: fix VBD.unplug where the frontend is in a managed domain (eg dom0)
Frontends in managed domains can guarantee to be unpluggable when
VBD.unplug is called (otherwise it's a bug in the management software)
Therefore it's safe and desirable to ignore the 'transient' "device
in use refusing to close" error.

Note this transient error has such a high false-positive rate that
it's not even that useful for unmanaged user domains. Probably it would
be better to rely on short timeouts and effective cancellation.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>