CA-82283: Always resync DHCP PIFs when xapi starts
The PIF.IP and PIF.netmask fields for PIFs that are configured by
DHCP are normally set by callbacks from the dhclient-script (via
xe host-signal-networking-change). When the hosts starts, this
call back comes before xapi is running. Therefore, we must resync
the IP fields whenever xapi starts, to be sure they properly reflect
the IP addresses of dom0. In fact, we must also resync when doing a
PIF.reconfigure_ip with mode=DHCP, in case the mode already was DHCP.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>