CA-80600: don't cancel our xenstore signal to qemu when a guest domain exits
The xenstore cancellation code cancels watches which are waiting for
a domain to respond, when the domain itself has exitted and will never respond.
In the case of qemu running in (eg) domain 0, we should cancel the watch
if domain 0 exits, not the guest domain.

This requires us to plumb through the correct qemu_domid eveywhere relevant.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>