CA-80781: When moving a VIF, only attach the associated network in Xapi_xenops
The original code attached the networks for all VIFs of the VM. However, vif_move is
typically called sequentially for several VIFs of a VM, e.g. when moving VIFs from a bond
slave to the bond master during a Bond.create. We should be careful not to try to attach
networks for bond slaves, and therefore attach only the network needed by the VIF that is
moved.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>