CA-82632: fix race in the squeeze deamon
This race happens when we are trying to get information from a domain which is just destroyed at the same time by xenops.

This fix doens't resolve the race, but it makes sure to catch the expression when something goes wrong and propagate it through the code.

Also, make the watch_xenstore event-loop more rebust by restarting it when an uncatched exception is thrown.