CA-83745: work-around a race between SR.create and PBD.plug by making PBD.plug more robust
The following interleaving appears to be possible:

XenAPI SR.create
  creates PBD
  adds SR to hashtable
  calls SMAPI sr_create
                                          XenAPI PBD.plug
                                            adds SR to hashtable
                                            sets currently_attached to true
  removes SR from hashtable
  calls XenAPI PBD.plug
    no-op since currently_attached = true

-- at this point a PBD is currently_attached = true but the SR is nolonger in the hashtable.

This race window always existed but it has recently been enlarged.

We work around this by unconditionally adding the SR to the hashtable in PBD.plug
irrespective of the value of currently_attached. Since SR.create always ends with
a call to PBD.plug, we gain the property that when the SR.create has finished
the SR must be in the hashtable.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>