CA-78139 - Make reading the inventory threadsafe.
If something (e.g. bringing up the management interface) is updating the
inventory, it's not safe to read from the hashtable.

This intermittently breaks the creation of the local SR at first boot,
since the sr-create command logs in to localhost which requires looking
up the installation uuid from the inventory.

Signed-off-by: John Else <john.else@citrix.com>