CA-83040: choose_suspend_sr ignores SRs not shared with host
When a VM was migrated within the pool from local storage to local storage, the
suspend SR was not reset, which caused suspend to fail because the VM's
suspend_SR was not visible to the host on which it was being suspended. We fix
the problem by modifying choose_suspend_sr to ignore a VM's suspend_SR field if
that SR has no PBD linking it with the VM's resident host.

Another commit will follow which blanks out the suspend_SR field on a VM when
it's suspend SR will no longer be visible after a migrate.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>