CA-85761: switch most UUID generation over to fast PRNG, rather than /dev/urandom
We only need /dev/{u,}random when the results have to be unguessable.
In the vast majority of cases we only require the UUIDs to be Unique.

When we need an unguessable UUID we use Uuid.secure() and when we need
only a unique UUID we use Uuid.insecure()

In particular, we still need unguessable UUIDs for anything to do with
* session generation
* secrets

Signed-off-by: David Scott <dave.scott@eu.citrix.com>