CA-75697: Disallow VBD.create during VM migration
All manner of bad things happen if a VBD is created during a VM's migration, and such new VBDs bypass the checks that would normally happen before a migrate can start.

This patch raises an "other operation in progress" error if you try to create a VBD on a VM that's in the middle of a migration.

Example output:

  # xe vbd-create vdi-uuid=<vdi> vm-uuid=efc5e680-21a1-b0b5-7787-a49fe4ebec34 device=autodetect
  Another operation involving the object is currently in progress
  class: VM
  object: efc5e680-21a1-b0b5-7787-a49fe4ebec34 (demo)
  <extra>: pool_migrate

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>