CA-84354: SXM fails when VM has snapshot
We weren't clearing the vdi_info.sm_config map when we created a new VDI on the
remote host for the snapshot VDI. The sm_config map still had the
host_OpaqueRef:... field, which caused blktap2.py to think that the VDI was
already attached and activated, when in fact it wasn't. We now clear the
sm_config map (although it is unclear whether clearing it is better than just
filtering the host_OpaqueRef key).

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>