CA-85940: add a test-case to check each "cancel point"
The test-case measures the number of "cancel points" traversed by
each lifecycle operation. It then checks that, if a "cancel point"
is triggered, that the system ends up getting into a valid state.

In particular it looks for:

 * xen domains and devices
 * xenopsd VM metadata
 * xapi VM metadata

all being in sync.

Run the test on an arbitrary test VM as follows (in dom0):

/root/cancel_tests -pw <password> -vm <VM name>

Signed-off-by: David Scott <dave.scott@eu.citrix.com>