CA-85940: add the concept of 'active' VBDs and VIFs to xenopsd
During a VM.start we set all VBDs and VIFs to active. Within
xenopsd a {VBD,VIF}.plug will be ignored if the device is not
active. This means that on resume, when we bring back exactly
those devices which were marked as "active" at the time of
the preceeding VM.start.

We now distinguish between a VBD_plug as part of a compound
operation and a user-requested VBD_hotplug, where the latter
also makes the device active.

We set XenAPI {VIF,VBD}.currently_attached = active || plugged
 -- this means that on suspend, currently_attached stays true
    since we never set active to "false"

Unfortunately the active flags get wiped in the middle of a
reboot, so a reboot behaves the same way as a start i.e. it
connects all devices, not just those which were connected before
the shutdown.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>