CA-86263: Revert patch for CA-85191.
The patch that's been reverted cancelled outstanding watches for devices
associated with a domain when the domain was destroyed. However, the watches
are for dom0 events which should not be dependent on the state of the
domU. In the case where the original problem was identified, the domU had
crashed and been left paused, leaving the devices up. The watches waiting
for the backends to go away timed out, which is actually not unreasonable
given that the backends still existed at that point.

Signed-off-by: Jon Ludlam <jonathan.ludlam@eu.citrix.com>