CA-86139: memory-actual still zero 2 mins after host became enabled
When master restarts, VM metrics for VMs on the slaves need to be updated.

This was one of the functionalities that the RRD code that was party of xapi
was previously performing in a fairly convoluted way. During the RRD code
refactoring, this functionality was lost. It is now implemented with a simple
cache-clear operation on the slave.

Signed-off-by: Rok Strniša <rok.strnisa@citrix.com>
Committed-by: Jon Ludlam <jonathan.ludlam@eu.citrix.com>