CA-85940: fix a race between deleting the guest metrics and updating it
If the VM exposes guest agent data very early then we will delete
the old guest metrics (because a reboot has happened) but fail to
notice the new values are present and fail to create the new guest metrics.

We now check the data from the guest agent immediately after we
spot a reboot and delete the old guest agent data.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>