CA-87459: if an event from xapi is received for a VM which has been destroyed, skip and don't pause for 15s
It's possible to have a VM be shutdown and uninstalled before all
events have been processed. If this happens then we log the
event and avoid calling "Thread.delay 15."

When a VM is actually shutdown, we trigger a xapi event re-registration
since we don't need to receive events on it any more.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>