CA-87552: add "Network.other_config:assume_network_is_shared=true" to override internal network checks
Normally we block VM.start, migrate etc if a VIF is on an "internal"
network which is already in use on another host. In the longer term we
should just tunnel, but in the short term we allow clients to specify
whether they want to allow the operations instead, per Network.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>