CA-86764: SXM intra-pool shouldn't move unmapped VDIs
On intra-pool SXM migrate, we shouldn't move unmapped VDIs, but should instead
leave them on their current SR. This was causing strange behavior on
VDI.pool_migrate, which would attempt to move single VDIs, but would find the
other VM's VDI(s) moving as well.

We should consider blocking a cross-pool migrate_send if the VDI map isn't
fully specified, instead of guessing and sending VDIs to the default SR. We'll
leave that for another commit.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>