CA-86764: Move Xapi_vdi.pool_migrate to Xapi_vm_migrate.vdi_pool_migrate.
To fix the deadlock mentioned in the previous commit, we needed to call
migrate_send directly, instead of going through the message forwarding layer.
This caused a mutually recursive module error, so we had to move
VDI.pool_migrate to Xapi_vm_migrate.

Note that we've moved the code which determines which host to run the local API
call on to *inside* the critical section, so that if the VM is moved by a
previous migration that we're waiting on, we don't try running this command on
the wrong host.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>