CA-88488: VM.reboot still exposed transient states where VBDs,VIFs were unplugged
A VBD or VIF is said to be 'active' while it is associated with a VM
even over a reboot. Unfortunately the record of which devices are 'active'
was stored in xenstore and deleted during the destroy phase of a reboot.

We now delay the deletion of the /vm/<uuid> tree until the VM metadata
is removed from xenopsd.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>