Revert the rebalance-interval changes to networkd of f03a30793ccdbf0b6071a5bfe5d8de06aa2984df
The policies for which values are set in the network layer should be in xapi. Networkd should
just do do what xapi says.

After this change, the default rebalance-intervals are unchanged: for balance-slb bonds is 30min,
while for LACP bonds, the value is unset and the default from the ovs is used (10s). However, the
change makes it possible for the user to override the rebalance-interval using the other-config key,
even in the LACP case (for whatever reason).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>