IO.open_connection now takes a Uri.t, rather than a concrete address
This will allow an app not to know how addresses are stored internally,
and allows xen-api clients to support different address types such as
* unix domain sockets -- never appropriate for mirage
* SSL -- future work for mirage

Of course the cost of this flexibility is the run-time error that
will be generated if the client doesn't support the Uri/address/scheme.
Given that runtime errors are usually possible anyway (eg DNS lookup
failures), the app has to be defensively-written anyway.