Add Unix/POSIX IO modules for Cohttp, both buffered and unbuffered
The buffered IO module uses the built-in ocaml in_channel/out_channel.

The unbuffered IO module only buffers the HTTP headers, and allows
the file descriptor to be passed, with the body unread, to another
process. Unfortunately this means we must issue more syscalls than we
would like (because we don't know the header length and can't just
do a big read()). We attempt to mitigate this by reading up to 4
bytes at a time, since the end of header marker (\r\n\r\n) is 4 bytes
long.

It might be better to always use a custom buffered channel implementation
and then always transmit the (remaining buffer, file descr) to
helper processes.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>