Hack to allow migrations during pool upgrade
Hack to allow XCP Boston hosts to migrate to XCP Tampa hosts during an upgrade.
If we are in rolling pool upgrade mode, we will add the key
product-version=$PLATFORM_VERSION to the software-version map. We start a
thread to watch for the pool to leave upgrade mode, and then remove the key. We
should only do this (as in, add the key and then remove it) if it is not already
set by the inventory file.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>