CA-90857: VIF Port Locking Persistence
Issue:
While importing a VM through XenCenter, the VIF imported is destroyed and recreated.
The locking-mode field for VIF was of type DynamicRO and the VIF recreated by XenCenter always used to set it to network default.

Solution:
The locking-mode field for VIF was changed to type StaticRO, the constructor signature modified and then the vif record was constructed locally to preserve this field.

Signed-off-by: Siddharth Vinothkumar <siddharth.vinothkumar@citrix.com>