If we can't determine the domU architecture, we assume it will be able to autonegotiate the block protocol correctly.
Very old 64-bit PV kernels had a bug where they would fail to negotiate
the 64-bit version of the block protocol. With a 32-bit domain 0 this
caused problems, so a workaround was added to the toolstack.

Recent xenctrl libraries have not got the function we need to determine
the domU architecture, so the workaround doesn't work.

If the workaround is unavailable we log and continue, rather than failing
to start the VM (which is almost certainlly going to work perfectly unless
it is very old)