CP-4091: add SM.features field to API
SM.features supercede SM.capabilites. Features represent versioned
capabilities. This will allow clients to predicate operations on a particular
verison of an SM capability, so they can be sure that a particular bug has been
fixed (which is the case with the VDI_RESET_ON_BOOT capability, which had
serious bugs in version 1).

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>