CA-83270: Remove the default value to verify_cert.
It prevents Stunnel.connect of actually checking the certificate.
The line: assert (not (verify_cert && use_tunnel_cache));
must be removed since verify_cert=None is undefined.
However, stunnel_cache need to be able to differentiate verified and
unverified connections after this commit, therefore the changes for
XOP-229 are necessary.

Signed-off-by: Jerome Maloberti <jerome.maloberti@citrix.com>
Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>