CA-101922: Properly close the nbd connection
The recent change to default http connections to v1.1 and
keep-alive=true was probably the cause of failures in storage
migration. This patch reverts the connection in question to
the previous behaviour

It's likely what happened was that xapi kept the connection
open after the fd was handed off to tapdisk and consumed
the requests coming in from the source.

Signed-off-by: Jon Ludlam <jonathan.ludlam@eu.citrix.com>