CP-4460: getter/setter/ctor/dtor etc. should inherit the lifecyle values from their corresponding fields/objects
so that when a field/object type is marked as removed, these implicit methods will reject clients with clear error message.

Note that, object/field/method types maked as removed will remain exist/effective in the DB layer, as they are also necceary in the case of upgrade (from previous releases).

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>