CP-4460: Add some nonconvential WLB "API" back to datamodel
There are not standard API messages, but web services and errors. They don't have a concept of API "lifecycle" like the standard ones, so we can't simply mark them as "deprecated" or "removed" and take advantage of the logic in our code autogeneration engine.

Instead of removing them, we have to keep the stubs around for the following reasons:

  * SDK auto generates some binding parts out of them (as well)
  * And, XenCenter will compile with a single version of API bindings and use it talk with different versions of XenServer (hence those binding stubs have to be a superset of all the historic versions we currently support).

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>