CA-104483: rrdd: ensure plugins' checksums change
Previously, rrdd only checked whether the plugin's checksum is different from the first one it saw for that plugin. This had the effect that, if the checksum changed at least once since rrdd was started, the data was always deemed to be fresh. The logic was flawed: Hashtbl.replace was only called when the entry wasn't previously in the hashtable.

Instead, this patch checks the current checksum against the most recently seen one. The effect of this is that data is only deemed fresh if it has been updated recently.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>