CA-105159: Wait for ack of shutdown request for HVM guests with PV drivers
The code paths only handled fully-PV guests and HVM guests without PV drivers.
In the case of PV drivers we did not wait for the shutdown request to be
acknowledged. This did not enable early exit if the request was not
acknowledged.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>
(cherry picked from commit 6de03e86b7804c63749d9ebd764839f66492d30d)